Optimize TitleFormatter::getPrefixedText() to Title performance
...by instance caching the result on the LinkTarget object.

Title::getPrefixedText() is an extremely hot function that has been
fully optimized over the years due to its usage basically everywhere in
MediaWiki.

Apply the same treatment to TitleFormatter, but keep the caching as part
of the LinkTarget instance so we can share caches across Title and
TitleFormatter. Once Title is switched over to using TitleFormatter
internally, we can probably centralize the cache again, and get rid of
the public but marked @private member variables.

Bug: T201801
Change-Id: I4ae2d6b176f69f66720473aeae85e39a601a9781