jquery.accessKeyLabel: Optimise by using mw.message().plain()
Avoids invoking mw.message#text, which uses the parser() of jqueryMsg,
twice for each call to jquery.accessKeyLabel.

These two messages do not contain any wikitext features, and are
not expected to need them in the future.

To make this explicit, add `{{format|plain}}` to their documentation,
and change the code to use plain().

Bug: T202357
Change-Id: I642f09afda4a93a1208550d44072f0909dcabdd0