resourceloader: Remove unused code from mw.loader.register()
The optional 'dependencies' parameter is only ever used with
an array of strings, never just as a string.

Also fix the missing square-brackets around the parameter name
in the documentation that indicate the parameter as optional.
The code was already allowing it to be missing, and the server
does indeed omit it when a module has no dependencies.
(See ResourceLoaderStartupModuleTest for confirmation of that).

The server-side methods ResourceLoaderModule::getDependencies
and ResourceLoader::makeLoaderRegisterScript already don't
support string-type for this parameter.

Change-Id: I80a1e4b2eefb62e669c031e0d953bf74a9623264