resourceloader: Remove redundant calls to getState() in mw.loader.enqueue()
The list of module names are known to be registered there.
Thus, no need for getState() and its hasOwn.call(registry) check.

If it was possible for an unknown module name to make its way to this code,
it would break both before and after this commit due to the setting of
registry[module].state which would throw TypeError for setting on undefined.

Change-Id: Ie3cccc7e4c57648279325852f362d5eb72128306