Special:Preferences: Split the unknown/neutral gender into label and help
The extensive label for the "how" of the form in English obscures the "what". By
splitting the radio button's label into a short "what" and retaining the current
message, which in English is currently *entirely* a "how", allows translators to
retain the opportunity to advise their users about the manner in which this will
affect their experience.

Change-Id: I5ca2fb56ee4b50bd6af9b13dfe3c50c1f661a4e8