user: Allow "CAS update failed" exceptions to be normalised
Take the user_id variable out of the exception message.
To compensate and still allow one to correlate patterns relating
to a specific user (e.g. a bot), add a warning message that
still contains the variable via context. This way that warning
will also be normalised/grouped, but with the extra context.

This is separate because exceptions do not currently support
context placeholders.

Bug: T202149
Change-Id: Ic0c25f66f23fdc65821da12f949c6224bc03f9b3