debug: Allow the DBQuery channel to be used
Follows-up 81e8d7af4129eb56 and e861191b2cdceeeecee.

When using $wgShowDebug, $wgDebugComments, or $wgDebugLogFile
locally, or when using the 'log' attribute with X-Wikimedia-Debug,
all channels should be enabled and logged. But, the DBQuery logs
are currently going nowhere.

The code for MWDebug::query() is intercepting it, even if
$wgDebugToolbar is not enabled.

And after that, the code for wfLogDBError was intercepting it,
again, even if $wgDBerrorLog is not enabled.

Bug: T202764
Change-Id: I710c26a9e9c30fea20975d1bc24e1f0af077c2ad