resourceloader: Use null-objects for 'sources' and 'registry'
Follows-up e5912535aea, and removes the last use of
Object.prototype-inheriting objects for map-like objects, in
mediawiki.js.

I'd like to consider using ES6 Map as well, with a partial shim
like we do for StringSet/Set, but that'll require some more
refactoring. It's also not clear whether it will improve
execution speed and/or memory use. Worth trying at a later time,
though.

The current change simply removes the inheritance and simplifies
the code. Apart from slightly smaller code, I could not find any
notable/reproducible improvement in either NavTiming metrics or
CPU time spent in 'Scripting'. The metrics are in the same range
as before this changes.

Change-Id: Ie0016667d9291dcfafde61289d5444817be3447d