Diff and history link separated via CSS
Piped tag list wrapped with parentheses taken care of via CSS.
Lists of links inside history are now a list of links rather than
2 anchors separated by text.

Remains translateable, and accessible but is now fully skinnable by 
skins such as Timeless and MinervaNeue.

For now the approach only touches the contributions pager, but
can be used on other pages as we work more on them.

Additional changes:
* A CSS file is renamed to a LESS file

Bug: T205581
Change-Id: Ic4a9c75611fd7ce92511c00c32b73c0745081b50