resources: Register mustache.js as foreign resource
The original commit that added it (7c52c9265b) did not specify
where it came from or which version. The source refers to
<https://github.com/janl/mustache.js> and contains a version string
stating "0.8.2". However, this file does not match the source
code of released version "0.8.2".

Instead, our copy seems to match the source ode of release
v1.0.0 (except for the version string).

The only commit in the uptream repository that matches our
copy is the commit right before the "Tag v1.0.0" commit,
at which point it had all changes of v1.0.0, but still had
the version string of the last published version. I assume then
at this file was copied from the GitHub master branch at the
time.

I'll update it to v1.0.0 proper in a separate commit.

Change-Id: I37c69e71fd4b005532cb0ac80964c72d3921eae6