Special:Preferences: Split up "Advanced options" on watchlist and RC tabs
On Special:Preferences, in the "Watchlist" and "Recent changes" tabs,
there is currently an "Advanced options" section, which contains a
huge list of mostly unrelated checkboxes.

Split off some of them that are actually related to two new sections,
"Watched pages" and "Changes shown", to make it easier to understand
their purpose.

Change-Id: I9bf91b78f6af49214a5d467d0896c63eb376d3d2